+++ src/main/java/roomescape/auth/service/WaitingAuthorizationService.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.auth.service;
 
 import java.util.Objects;
 import org.springframework.stereotype.Service;
