+++ src/main/java/roomescape/controller/manager/ManagerReservationController.java
@@ -11,12 +11,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.LoginMember;
+import roomescape.auth.service.ReservationAuthorizationService;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.dto.request.ReservationPatchDto;
 import roomescape.dto.response.AdminReservationResponseDto;
 import roomescape.service.AdminReservationService;
-import roomescape.service.ReservationAuthorizationService;
 
 @RestController
 @RequestMapping("/manager/reservations")
