+++ src/test/java/roomescape/controller/manager/ManagerWaitingControllerTest.java
@@ -30,7 +30,7 @@ import roomescape.domain.Time;
 import roomescape.domain.Waiting;
 import roomescape.domain.vo.Name;
 import roomescape.dto.response.WaitingResponse;
-import roomescape.service.WaitingAuthorizationService;
+import roomescape.auth.service.WaitingAuthorizationService;
 import roomescape.service.WaitingService;
 
 @WebMvcTest(ManagerWaitingController.class)
