+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -1,5 +1,6 @@
 package roomescape.service;
 
+import roomescape.auth.service.ReservationAuthorizationService;
 import roomescape.common.exception.BusinessRuleViolationException;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -17,7 +18,6 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.common.exception.InvalidInputException;
 import roomescape.common.exception.DuplicateEntityException;
 import roomescape.common.exception.EntityNotFoundException;
 import roomescape.common.exception.HiddenResourceException;
